+++ modules/governance/events/outputs.tf
@@ -15,6 +15,51 @@ output "cis_level" {
 }
 
 output "resource_ids" {
-  description = "Map of resource identifiers created by this module. Empty until implementation."
-  value       = {}
+  description = "Map of resource identifiers created by this module."
+  value = merge(
+    {
+      for key, topic in oci_ons_notification_topic.this : "topic.${key}" => topic.id
+    },
+    {
+      for key, subscription in oci_ons_subscription.this : "subscription.${key}" => subscription.id
+    },
+    {
+      for key, rule in oci_events_rule.this : "event_rule.${key}" => rule.id
+    }
+  )
+}
+
+output "notification_topic_ids" {
+  description = "Map of notification topic keys to OCIDs."
+  value = {
+    for key, topic in oci_ons_notification_topic.this : key => topic.id
+  }
+}
+
+output "notification_topic_names" {
+  description = "Map of notification topic keys to names."
+  value = {
+    for key, topic in oci_ons_notification_topic.this : key => topic.name
+  }
+}
+
+output "subscription_ids" {
+  description = "Map of subscription keys to OCIDs."
+  value = {
+    for key, subscription in oci_ons_subscription.this : key => subscription.id
+  }
+}
+
+output "event_rule_ids" {
+  description = "Map of event rule keys to OCIDs."
+  value = {
+    for key, rule in oci_events_rule.this : key => rule.id
+  }
+}
+
+output "event_rule_names" {
+  description = "Map of event rule keys to display names."
+  value = {
+    for key, rule in oci_events_rule.this : key => rule.display_name
+  }
 }